+++ tests/testthat.R
@@ -1,4 +1,5 @@
-library(testthat)
+suppressWarnings({library(testthat)
 library(zooper)
+})
 
 test_check("zooper")