+++ tests/testthat/test-Data.R
@@ -1,7 +1,9 @@
-library(zooper)
-require(dplyr)
-require(stringr)
-require(lubridate)
+suppressWarnings({
+  library(zooper)
+  require(dplyr)
+  require(stringr)
+  require(lubridate)
+})
 
 Data_sets <- c("EMP_Meso", "FMWT_Meso", "STN_Meso",
                "20mm_Meso", "FRP_Meso","EMP_Micro",